Refactor routes Implement lombok
