+++ src/main/java/org/tech/mycvtech/model/user/User.java
@@ -1,14 +1,20 @@
-package org.tech.mycvtech.model;
+package org.tech.mycvtech.model.user;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.tech.mycvtech.model.util.Rol;
 
 import java.util.Date;
 
 @Document("Users")
+@Builder
+@Data
+@AllArgsConstructor
 public class User {
     @Id
     protected String id;
@@ -25,66 +31,14 @@ public class User {
         this.createdAt = new Date();
     }
 
-    public String getPhone() {
-        return phone;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
     @JsonIgnore
     public String getPassword() {
         return password;
     }
 
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public Rol getRol() {
-        return rol;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
     @JsonProperty
     public void setPassword(String password) {
         this.password = password;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
     }
 
     protected void update(User userUpdates) {
