+++ src/main/java/org/tech/mycvtech/model/util/Knowledge.java
@@ -1,22 +1,21 @@
-package org.tech.mycvtech.model;
+package org.tech.mycvtech.model.util;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class Knowledge {
     @JsonProperty("name")
     String name;
     @JsonProperty("level")
     Level level;
 
-    public Knowledge(String name, Level level) {
-        this.name = name;
-        this.level = level;
-    }
-
     public Knowledge(String name) {
         this(name, Level.ZERO);
     }
 
-    public Knowledge() {
-    }
 }
