+++ src/main/java/org/tech/mycvtech/model/util/SoftSkill.java
@@ -1,0 +1,15 @@
+package org.tech.mycvtech.model.util;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SoftSkill {
+    @JsonProperty("name")
+    String name;
+}
