+++ src/main/java/org/tech/mycvtech/repository/UserMongoRepository.java
@@ -2,7 +2,7 @@ package org.tech.mycvtech.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-import org.tech.mycvtech.model.User;
+import org.tech.mycvtech.model.user.User;
 
 @Repository
 public interface UserMongoRepository extends MongoRepository<User, String> {
